requirements.txt
marvin
openai
halo
PyYAML
